Further fix for dataSize bug
Further correction was required for the readFromFlash function to accommodate the Long data type

+++ SPIFlash/SPIFlash.cpp
@@ -385,7 +385,7 @@ char SPIFlash::readFromFlash(int pageAddr, int byteAddr) {
 }
 
 
-char SPIFlash::readFromFlash(int byteAbsAddr) {
+char SPIFlash::readFromFlash(unsigned long byteAbsAddr) {
 
   char buf[10];
   int temp;
